requirements.txt
selenium
flake8
ipython
pytest
allure-pytest